Much faster sorting with internal @id attribute + a lot of bug fixes
Compressed document id data can now be read and decompressed in both
directions - ascending and descending. This enables the PMBApi to choose
an appropriate direction when sorting by the internal @id attribute and
simply stop the decompression when a suitable amount of results has been
found. Huge performance improvements ( +1000% ) in selected situations.

Now, there is, however, a downside using the internal @id attribute for
sorting: the exact amount of results is unknown.  Based on the density (
actual matches / decompressed values ) of results PMBApi attempts to
calculate an estimated number, but more often than not it is inaccurate.
Only way to know for sure would be to decompress all the data, like
before. For implementing this feature correctly please see the updated
PMBApi instructions on our website.

This release contains also numerous bug fixes. Updating is certainly
recommended.

Unfortunately the index data format has been also changed, so
re-indexing is required. If you are using the internal @id attribute for
sorting results descending order in your application, please refrain
from replacing the PMBApi.php file with the updated version until you've
had a change to re-index.